requirements.txt
streamlit>=1.30.0
pandas>=2.1.0
python-dotenv>=1.0.0
google-generativeai>=0.5.0
plotly>=5.16.0
psycopg2-binary
langfuse==2.25.0
